Refactor the structure of packages to follow the naming convention
