--- src/main/java/cs2103t/duke/exception/DukeException.java
@@ -1,9 +1,0 @@
-package cs2103t.duke.exception;
-
-public class DukeException extends RuntimeException {
-
-    public DukeException(String message) {
-        super(message);
-    }
-
-}
