--- src/main/java/cs2103t/duke/task/TaskType.java
@@ -1,7 +1,0 @@
-package cs2103t.duke.task;
-
-public enum TaskType {
-    TODO,
-    DEADLINE,
-    EVENT,
-}
