+++ src/main/java/fan/cs2103t/duke/exception/DukeException.java
@@ -1,0 +1,9 @@
+package fan.cs2103t.duke.exception;
+
+public class DukeException extends RuntimeException {
+
+    public DukeException(String message) {
+        super(message);
+    }
+
+}
