+++ src/main/java/fan/cs2103t/duke/task/TaskType.java
@@ -1,0 +1,7 @@
+package fan.cs2103t.duke.task;
+
+public enum TaskType {
+    TODO,
+    DEADLINE,
+    EVENT,
+}
