+++ src/test/java/fan/cs2103t/duke/parser/ParserTest.java
@@ -1,0 +1,23 @@
+package fan.cs2103t.duke.parser;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.Test;
+
+import fan.cs2103t.duke.ui.Ui;
+
+class ParserTest {
+
+    @Test
+    public void parseCommand_wrongDeadlineFormat_exceptionThrown() {
+        String space = Ui.SPACE;
+        try {
+            new Parser().parseCommand("deadline read book /by tomorrow");
+            fail();
+        } catch (Exception e) {
+            assertEquals(space + "OOPS!!! The deadline is in the wrong format.", e.getMessage());
+        }
+    }
+
+}
